Add new option and rewrite loader functionality
