+++ src/plugin.js
@@ -1,3 +1,7 @@
+const fs = require('fs');
+const XXHash = require('xxhash');
+
+const loaderPath = require.resolve('./loader.js');
 const { createSprite } = require('./spriteUtils');
 
 const BODY_TAG_BEGIN = '<body';
@@ -16,12 +20,21 @@ module.exports = class SvgSpriteHtmlWebpackPlugin {
   constructor(options = {}) {
     this.nextSymbolId = 0; // use only by this.generateId
     this.generateSymbolId = options.generateSymbolId || this.generateSymbolId.bind(this);
+    this.includeFiles = options.includeFiles || [];
     this.svgList = [];
     this.lastCompiledList = this.svgList;
     this.svgSprite = '';
 
+    this.pushSVGIncomingFile = this.pushSVGIncomingFile.bind(this);
     this.pushSvg = this.pushSvg.bind(this);
     this.processSvg = this.processSvg.bind(this);
+
+    if (this.includeFiles.length) {
+      this.includeFiles.map(e => {
+        fs.readFile(e, (err, data) =>
+          this.pushSVGIncomingFile(e, data, data.toString('utf-8')));
+      });
+    }
   }
 
   /**
@@ -78,6 +91,7 @@ module.exports = class SvgSpriteHtmlWebpackPlugin {
         // Svg loader will push imported svg paths and ids
         if (!loaderContext.svgList || !loaderContext.generateSymbolId) {
           loaderContext.pushSvg = this.pushSvg;
+          loaderContext.pushSVGIncomingFile = this.pushSVGIncomingFile;
           loaderContext.generateSymbolId = this.generateSymbolId;
         }
       });
@@ -139,10 +153,36 @@ module.exports = class SvgSpriteHtmlWebpackPlugin {
   }
 
   /**
+   * Function called when HTML string is generated by HtmlWebpackPlugin
+   * This function generate svg sprite with the list of svg collected by the svgLoader (./loader.js)
+   * When the sprite is generated, it's injected in the HTML string
+   * @param {string} filePath - represents the system path to the svg file
+   * @param {Buffer} fileBuffer - a Node Buffer object from the file used inside the caching function
+   * @param {string} fileContent - a string representation of the svg object
+   */
+  pushSVGIncomingFile(filePath, fileBuffer, fileContent) {
+    const svgPath = filePath;
+    const svgHash = XXHash.hash(fileBuffer, 0xCAFEBABE);
+    const symbolId = this.generateSymbolId(svgPath, svgHash, fileContent);
+
+    const svgItem = {
+      id: symbolId,
+      hash: svgHash,
+      path: svgPath,
+      content: fileContent,
+    };
+
+    // pushSvg is injected by the plugin SvgSpriteHtmlWebpackPlugin (./plugin)
+    this.pushSvg(svgItem);
+
+    return `export default '#${symbolId}'`;
+  }
+
+  /**
    * Resolve the path of webpack loader to add in webpack configuration
    * @return {string} - the path of the svg loader
    */
   static getLoader() {
-    return require.resolve('./loader.js');
+    return loaderPath;
   }
 };
